Basic processes completed for inserting single item, updating single item, deleting single item. Plus made some changes so integer can be used instead of long for getting id back. I always used id with no problem. If I ever run into a problem where used too many ids, then would require rethinking (everything turning to long. Next step is making public extensions not use interface but something else to help with intellisense experience when using.

+++ SqlExtensions/InsertSimple.cs
@@ -11,24 +11,24 @@ public static class InsertSimple
         PopulateSimple(map.Columns, output, EnumCategory.UseDatabaseMapping);
         return output;
     }
-    public static long InsertSingle<E>(this ICaptureCommandParameter connector, E thisObject, IDbTransaction? thisTran = null, int? connectionTimeOut = null) where E : class, ISimpleDapperEntity
+    public static int InsertSingle<E>(this ICaptureCommandParameter connector, E thisObject, IDbTransaction? thisTran = null, int? connectionTimeOut = null) where E : class, ISimpleDapperEntity
     {
         EnumDatabaseCategory category = connector.Category;
         CompleteSqlData data = PrivateGetInsert(thisObject, category, out bool IsAutoIncremented);
-        long id = connector.ExecuteScalar<long>(data, thisTran, connectionTimeOut); //this will focus on text alone.  may later think about stored procedures
+        int id = connector.ExecuteScalar<int>(data, thisTran, connectionTimeOut); //this will focus on text alone.  may later think about stored procedures
         if (IsAutoIncremented == true)
         {
             return id;
         }
         return thisObject.ID;
     }
-    public static async Task<long> InsertSingleAsync<E>(this ICaptureCommandParameter connector, E thisObject, IDbTransaction? thisTran = null, int? connectionTimeOut = null) where E : class, ISimpleDapperEntity
+    public static async Task<int> InsertSingleAsync<E>(this ICaptureCommandParameter connector, E thisObject, IDbTransaction? thisTran = null, int? connectionTimeOut = null) where E : class, ISimpleDapperEntity
     {
         EnumDatabaseCategory category = connector.Category;
         CompleteSqlData data = PrivateGetInsert(thisObject, category, out bool IsAutoIncremented);
         //for now, will always focus on text (not stored procedures).
 
-        long id = await connector.ExecuteScalarAsync<long>(data, thisTran, connectionTimeOut); //this will focus on text alone.  may later think about stored procedures
+        int id = await connector.ExecuteScalarAsync<int>(data, thisTran, connectionTimeOut); //this will focus on text alone.  may later think about stored procedures
         if (IsAutoIncremented == true)
         {
             return id;
